bin/requirements.txt
pyserial
mock

